Added support for zip (per scan line) compression
This commit also resolves huge Nuke problem. If there was specified zips as a
compression, data was stored in zip (per 16 lines) format and Line Offset Table
was huge but with correct data. Sometimes it was a big problem for Nuke because
Nuke reads raw data using Imf::InputFile::rawPixelData